Add commander, command line script dispatcher
