+++ package.json
@@ -80,6 +80,7 @@
     "@0xproject/sol-compiler": "^0.5.2",
     "@parity/api": "^2.1.20",
     "bignumber.js": "^4.1.0",
+    "commander": "^2.17.1",
     "ethereumjs-testrpc": "^4.1.3",
     "ganache-cli": "^6.1.4",
     "mkdirp": "^0.5.1",
