+++ utils/kyber/populateDevConfig.js
@@ -6,7 +6,6 @@ const devchainConfigFile = "./utils/kyber/devchain-reserve.json";
 async function populateDevConfig() {
   const a = JSON.parse(fs.readFileSync(devchainConfigFile));
   const b = JSON.parse(fs.readFileSync(addressBookFile));
-  /* eslint-disable no-restricted-syntax */
   for (const i of Object.keys(a.tokens)) {
       a.tokens[i].address = b.development[i];
   }
